+++ packages/core/docs/changelog/5620.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'removed helpers/filters from @vue-storefront/nuxt-theme and @vue-storefront/commercetools-theme because they are useless now',
-  link: 'https://github.com/vuestorefront/vue-storefront/issues/5620',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'Kacper Małkowski',
-  linkToGitHubAccount: 'https://github.com/porithe/'
-};