+++ packages/core/docs/changelog/5693.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'extend #5692 thrown error to fix also this issue',
-  link: 'https://github.com/vuestorefront/vue-storefront/issues/5693',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'Baroshem',
-  linkToGitHubAccount: 'https://github.com/Baroshem'
-};