+++ packages/core/docs/commercetools/changelog/5708.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'Fixed mapping for product attributes with type `set`',
-  link: 'https://github.com/vuestorefront/vue-storefront/pull/5708',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'Filip Sobol',
-  linkToGitHubAccount: 'https://github.com/filipsobol'
-};